pyproject.toml
[tool.poetry]
name = "metadata"
version = "0.1.0"
description = ""
authors = ["diogenes69 <diogenes_monegro@claro.com.do>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
python-pptx = "^1.0.2"
openpyxl = "^3.1.5"
python-docx = "^1.1.2"
pytest = "^8.3.3"
black = "^24.10.0"
pytest-html = "^4.1.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
